Add ability to check all possible install states
This adds a relatively simple loop and switch to check the install state
of dragonfly. There are three things that are optionally checked with
their corresponding key (in parenthesis):

1) Dragonfly can be imported. (sanity)
2) Direct can be imported. (direct)
3) Optimal transport can be imported. (optimal)

If any of these are unable to import and the corresponding key is given
as an argument to the tests, then the tests will fail and print the
value that failed.

Example:

$ bash run_all_tests.sh sanity
This checks to make sure that dragonfly can be imported. This will
return a nonzero exit code if it can not.

$ bash run_all_tests.sh sanity direct
This will check that dragonfly and direct can be imported.

$ bash run_all_tests.sh sanity optimal
This will check that dragonfly and optimal transport can be imported.

Note: All of these tests come with the prerequisite that 'sanity' is
always the first test. This is because the others will not make sense
outside of that context.